Add connection ID for connections, use it for sorting
Closes #27
