+++ .gitignore
@@ -4,4 +4,5 @@ gradle-app.setting
 .idea/
 /out/
 /conf/
-config.yml
+/config.yml
+/stylesheet.css
